+++ CHANGES.rst
@@ -8,6 +8,8 @@ Changelog
   catalog attribute.
   [cekk]
 
+- Added ignore_portal_types and only_portal_types parameter to reindex for maintenance_view [jkubaile]
+
 
 4.0.3 (2014-06-18)
 ------------------